+++ pkg/morph/client/notary.go
@@ -30,9 +30,8 @@ import (
 
 type (
 	notaryInfo struct {
-		txValidTime  uint32 // minimum amount of blocks when mainTx will be valid
-		roundTime    uint32 // extra amount of blocks to synchronize sidechain height diff of inner ring nodes
-		fallbackTime uint32 // mainTx's ValidUntilBlock - fallbackTime + 1 is when fallbackTx is sent
+		txValidTime uint32 // minimum amount of blocks when mainTx will be valid
+		roundTime   uint32 // extra amount of blocks to synchronize sidechain height diff of inner ring nodes
 
 		alphabetSource AlphabetKeys // source of alphabet node keys to prepare witness
 
@@ -43,7 +42,7 @@ type (
 	notaryCfg struct {
 		proxy util.Uint160
 
-		txValidTime, roundTime, fallbackTime uint32
+		txValidTime, roundTime uint32
 
 		alphabetSource AlphabetKeys
 	}
@@ -53,9 +52,8 @@ type (
 )
 
 const (
-	defaultNotaryValidTime    = 50
-	defaultNotaryRoundTime    = 100
-	defaultNotaryFallbackTime = 40
+	defaultNotaryValidTime = 50
+	defaultNotaryRoundTime = 100
 
 	notaryBalanceOfMethod    = "balanceOf"
 	notaryExpirationOfMethod = "expirationOf"
@@ -71,7 +69,6 @@ func defaultNotaryConfig(c *Client) *notaryCfg {
 	return &notaryCfg{
 		txValidTime:    defaultNotaryValidTime,
 		roundTime:      defaultNotaryRoundTime,
-		fallbackTime:   defaultNotaryFallbackTime,
 		alphabetSource: c.Committee,
 	}
 }
@@ -106,7 +103,6 @@ func (c *Client) EnableNotarySupport(opts ...NotaryOption) error {
 		proxy:          cfg.proxy,
 		txValidTime:    cfg.txValidTime,
 		roundTime:      cfg.roundTime,
-		fallbackTime:   cfg.fallbackTime,
 		alphabetSource: cfg.alphabetSource,
 		notary:         notary.Hash,
 	}
@@ -671,15 +667,6 @@ func WithTxValidTime(t uint32) NotaryOption {
 func WithRoundTime(t uint32) NotaryOption {
 	return func(c *notaryCfg) {
 		c.roundTime = t
-	}
-}
-
-// WithFallbackTime returns a notary support option for client
-// that specifies amount of blocks before fallbackTx will be sent.
-// Should be less than TxValidTime.
-func WithFallbackTime(t uint32) NotaryOption {
-	return func(c *notaryCfg) {
-		c.fallbackTime = t
 	}
 }